// source/src/shared/lib/enums/actions/MainExampleAction.ts
export enum MainExampleAction {
    INSERT_ELEMENT_IN_EXAMPLE = 'INSERT_ELEMENT_IN_EXAMPLE',
    INSERT_RESULT_IN_EXAMPLE = 'INSERT_RESULT_IN_EXAMPLE',

    INSERT_CURSOR_IN_EXAMPLE_BY_PLACE = 'INSERT_CURSOR_IN_EXAMPLE_BY_PLACE',
    INSERT_CURSOR_IN_EXAMPLE_BY_INDEX = 'INSERT_CURSOR_IN_EXAMPLE_BY_INDEX',
    INSERT_CURSOR_IN_START_OF_EXAMPLE = 'INSERT_CURSOR_IN_START_OF_EXAMPLE',
    INSERT_CURSOR_IN_END_OF_EXAMPLE   = 'INSERT_CURSOR_IN_END_OF_EXAMPLE',
    
    INSERT_ELEMENT_IN_EXAMPLE_BY_INDEX = 'INSERT_ELEMENT_IN_EXAMPLE_BY_INDEX',
    DELETE_ELEMENT_IN_EXAMPLE_BY_INDEX = 'DELETE_ELEMENT_IN_EXAMPLE_BY_INDEX',
    DELETE_ELEMENT_IN_EXAMPLE = 'DELETE_ELEMENT_IN_EXAMPLE',
    DELETE_EXAMPLE = 'DELETE_EXAMPLE',
    ADD_EXAMPLE = 'ADD_EXAMPLE'
}